requirements.txt
arxiv~=1.4.7
discord==2.3.1
fastapi~=0.98.0
httpx~=0.24.1
motor~=3.2.0
openai~=0.27.8
pandas~=2.0.2
streamlit==1.24.0
strawberry-graphql[fastapi]==0.190.0
uvicorn[standard]~=0.22.0
python-dotenv
